requirements.txt
pytesseract
pillow
openai
flask
streamlit
groq
